requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
black==24.2.0
    # via -r requirements.in
blinker==1.7.0
    # via flask
click==8.1.7
    # via
    #   black
    #   flask
colorama==0.4.6
    # via click
flask==3.0.2
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
flask-sqlalchemy==3.1.1
    # via -r requirements.in
greenlet==3.0.3
    # via sqlalchemy
itsdangerous==2.1.2
    # via flask
jinja2==3.1.3
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
mypy-extensions==1.0.0
    # via black
packaging==24.0
    # via black
pathspec==0.12.1
    # via black
platformdirs==4.2.0
    # via black
psycopg==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
sqlalchemy==2.0.28
    # via
    #   -r requirements.in
    #   flask-sqlalchemy
    #   sqlalchemy-utils
sqlalchemy-utils==0.41.1
    # via -r requirements.in
typing-extensions==4.10.0
    # via
    #   psycopg
    #   sqlalchemy
tzdata==2024.1
    # via psycopg
werkzeug==3.0.1
    # via flask
